Fix task assignments update endpoint

+++ api/src/modules/task/task-assignment/task-assignment.service.ts
@@ -5,6 +5,7 @@ import { ApiHttpException } from 'src/exception/ApiHttpException.type';
 import { UnitOfWorkService } from 'src/modules/unit-of-work/unit-of-work.service';
 import { UpdateTaskAssignmentsStatusesRequest } from 'src/modules/workspace/workspace-module/dto/request/update-task-assignments-statuses-request.dto';
 import { ProgressStatus } from '../task-module/domain/progress-status.enum';
+import { TaskService } from '../task-module/task.service';
 import { TaskAssignmentCore } from './domain/task-assignment-core.domain';
 import { TaskAssignmentWithAssigneeUser } from './domain/task-assignment-with-assignee-user.domain';
 import { TaskAssignment } from './domain/task-assignment.domain';
@@ -15,6 +16,7 @@ export class TaskAssignmentService {
   constructor(
     private readonly taskAssignmentRepository: TaskAssignmentRepository,
     private readonly unitOfWorkService: UnitOfWorkService,
+    private readonly taskService: TaskService,
   ) {}
 
   async getAccumulatedPointsForWorkspaceUser({
@@ -90,38 +92,89 @@ export class TaskAssignmentService {
 
   async updateAssignessByTaskId({
     taskId,
+    workspaceId,
     data,
   }: {
     taskId: TaskAssignment['task']['id'];
-    data: UpdateTaskAssignmentsStatusesRequest;
+    workspaceId: TaskAssignment['task']['workspace']['id'];
+    data: UpdateTaskAssignmentsStatusesRequest['assignments'];
   }): Promise<TaskAssignmentWithAssigneeUser[]> {
+    // 1. Check that the given task exists in the given workspace
+    const task = await this.taskService.findByTaskIdAndWorkspaceId({
+      taskId,
+      workspaceId,
+    });
+
+    if (!task) {
+      throw new ApiHttpException(
+        {
+          code: ApiErrorCode.INVALID_PAYLOAD,
+        },
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
+    // 2. Get task assignments
     const taskAssignments = await this.findAllByTaskIdWithAssigneeUser(taskId);
 
-    const existingAssigneeIds = taskAssignments.map(
-      (taskAssignment) => taskAssignment.assignee.id,
+    if (taskAssignments.length === 0) {
+      // If there are no assignments, return empty array
+      return [];
+    }
+
+    const existingAssigneeMap = new Map(
+      taskAssignments.map((taskAssignment) => [
+        taskAssignment.assignee.id,
+        taskAssignment,
+      ]),
     );
-    const updatedAssigneeIds = data.map((data) => data.assigneeId);
-    const assigneeIdsToRemove = existingAssigneeIds.filter(
-      (existingAssigneeId) => !updatedAssigneeIds.includes(existingAssigneeId),
-    );
-    const assigneeIdsToAdd = updatedAssigneeIds.filter(
-      (existingAssigneeId) => !existingAssigneeIds.includes(existingAssigneeId),
+    const updatedAssigneeMap = new Map(
+      data.map((item) => [item.assigneeId, item]),
     );
 
-    await this.unitOfWorkService.withTransaction(() => {
+    const assigneeIdsToUpdate = Array.from(existingAssigneeMap.keys()).filter(
+      (id) => updatedAssigneeMap.has(id),
+    );
+    const assigneeIdsToRemove = Array.from(existingAssigneeMap.keys()).filter(
+      (existingAssigneeId) => !updatedAssigneeMap.has(existingAssigneeId),
+    );
+    const assigneeIdsToAdd = Array.from(updatedAssigneeMap.keys()).filter(
+      (updatedAssigneeId) => !existingAssigneeMap.has(updatedAssigneeId),
+    );
+
+    await this.unitOfWorkService.withTransaction(async () => {
       // 1. Delete task assignments if some assignees were removed from the task
-      this.taskAssignmentRepository.deleteByTaskIdAndAssigneeIds({
-        taskId,
-        assigneeIds: assigneeIdsToRemove,
-      });
+      if (assigneeIdsToRemove.length > 0) {
+        await this.taskAssignmentRepository.deleteByTaskIdAndAssigneeIds({
+          taskId,
+          assigneeIds: assigneeIdsToRemove,
+        });
+      }
 
       // 2. Create new task assignments if there are new assignees
       for (const assigneeId of assigneeIdsToAdd) {
-        this.create({
+        await this.create({
           workspaceUserId: assigneeId,
           taskId,
           status: ProgressStatus.IN_PROGRESS,
         });
+      }
+
+      // 3. Update existing task assignments statuses if needed
+      for (const assigneeId of assigneeIdsToUpdate) {
+        const existingAssignment = existingAssigneeMap.get(assigneeId);
+        const newStatus = updatedAssigneeMap.get(assigneeId)!.status;
+
+        if (
+          existingAssignment &&
+          newStatus &&
+          existingAssignment.status !== newStatus
+        ) {
+          await this.taskAssignmentRepository.update({
+            id: existingAssignment.id,
+            data: { status: newStatus },
+          });
+        }
       }
     });
